abstract drawing so devices may have full or half resolution; rebuild BIOS interface

+++ globals.cpp
@@ -9,7 +9,7 @@ PhysicalSpeaker *g_speaker = NULL;
 PhysicalPaddles *g_paddles = NULL;
 PhysicalPrinter *g_printer = NULL;
 VMui *g_ui;
-int16_t g_volume = 15;
+int8_t g_volume = 15;
 uint8_t g_displayType = 3; // FIXME m_perfectcolor
 VMRam g_ram;
 volatile bool g_inInterrupt = false;
